oauth access token acquistion process completed, started IoC consideration for comment-commenting feautre

+++ ConvApp/ConvApp/Models/CommentModel.cs
@@ -6,11 +6,11 @@ namespace ConvApp.Models
 {
     public class CommentModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public DateTime CreatedDate { get; set; }
         public byte ParentType { get; set; }
-        public int ParentId { get; set; }
-        public int UserId { get; set; }
+        public long ParentId { get; set; }
+        public long UserId { get; set; }
         public string Text { get; set; }
 
         public static async Task<CommentViewModel> Populate(CommentModel model)
