+++ ConvApp/ConvApp/Models/LikeModel.cs
@@ -5,10 +5,10 @@ namespace ConvApp.Models
 {
     public class LikeModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public byte ParentType { get; set; }
-        public int ParentId { get; set; }
-        public int UserId { get; set; }
+        public long ParentId { get; set; }
+        public long UserId { get; set; }
         public DateTime CreatedDate { get; set; }
 
         public static async Task<Like> Populate(LikeModel model)
@@ -23,7 +23,7 @@ namespace ConvApp.Models
 
     public class Like
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public UserBriefModel Creator { get; set; }
         public DateTime CreatedDate { get; set; }
     }
