+++ ConvApp/ConvApp/Models/ProductModel.cs
@@ -13,17 +13,17 @@ namespace ConvApp.Models
         public static string[] StoreNames = { "", "GS25", "CU", "미니스톱", "세븐일레븐", "이마트24", "홈플러스 익스프레스" };
         public static string[] CategoryNames = { "", "도시락", "라면", "김밥", "즉석식품", "스낵", "음료", "주류", "생활용품", "기타" };
 
-        public int Id { get; set; }
-        public int StoreId { get; set; }
+        public long Id { get; set; }
+        public long StoreId { get; set; }
         public string StoreName { get => StoreNames[this.StoreId]; }
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public string CategoryName { get => CategoryNames[this.CategoryId]; }
 
         public DateTime CreatedDate { get; set; }
 
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Price { get; set; }
+        public long Price { get; set; }
 
         public string Image
         {
